:recycle: refactor: 重构整合 createCSS 相关的方法，优化代码逻辑
