+++ src/core/index.ts
@@ -1,4 +1,3 @@
-export * from './createClassNameGenerator';
-export * from './createCX';
+export * from './createCSS';
 export * from './createEmotion';
 export { serializeCSS, type SerializeCSS } from './createSerializeStyles';
